크리에이터 confirm 수정함

+++ Class_Will/src/main/webapp/WEB-INF/views/creator/creator-classModify.jsp
@@ -357,7 +357,9 @@
 	<script type="text/javascript">	
 	
 		function DeleteClass(class_code) {
-			location.href="DeleteClass?class_code=" + class_code;
+			if(confirm("정말 삭제하시겠습니까?")){
+				location.href="DeleteClass?class_code=" + class_code;
+			}
 		}
 	
 		window.addEventListener('load', () => {
